move world scale out of preferences
makes auto default
