+++ microscopynodes/handle_blender_structs/units.py
@@ -12,7 +12,6 @@ DEFAULT_IMPORT_SCALE = "MICROMETER_SCALE"
 AUTO_TARGET_EXTENT_METERS = 10
 
 IMPORT_SCALE_ITEMS = [
-    (AUTO_IMPORT_SCALE, "(auto)", "Chooses a scale from loaded datasets so they fit within a few Blender meters", "", -1),
     ("NANOMETER_SCALE", "nm -> m", "Scales to 1 nm/blender-meter", "", 0),
     ("NANOMETER_DECIMETER_SCALE", "nm -> dm", "Scales to 1 nm/blender-decimeter", "", 1),
     ("NANOMETER_CENTIMETER_SCALE", "nm -> cm (Molecular Nodes)", "Scales to 1 nm/blender-centimeter", "", 2),
@@ -25,6 +24,7 @@ IMPORT_SCALE_ITEMS = [
     ("METER_SCALE", "m -> m", "Scales to 1 m/blender-meter", "", 9),
     ("METER_DECIMETER_SCALE", "m -> dm", "Scales to 1 m/blender-decimeter", "", 10),
     ("METER_CENTIMETER_SCALE", "m -> cm", "Scales to 1 m/blender-centimeter", "", 11),
+    (AUTO_IMPORT_SCALE, "(auto)", "Chooses a scale from loaded datasets so they fit within a few Blender meters", "", -1),
 ]
 
 IMPORT_SCALE_OUTPUT_SCALES = {
@@ -45,6 +45,82 @@ IMPORT_SCALE_OUTPUT_SCALES = {
 
 def import_scale_items():
     return list(IMPORT_SCALE_ITEMS)
+
+
+def _holder_extent_units(holder):
+    from mathutils import Vector
+
+    corners = []
+    for child in holder.children:
+        child_corners = getattr(child, "bound_box", None)
+        if not child_corners:
+            continue
+        for corner in child_corners:
+            corners.append(child.matrix_local @ Vector(corner))
+
+    if not corners:
+        return 0.0
+
+    return max(
+        max(corner[axis] for corner in corners) - min(corner[axis] for corner in corners)
+        for axis in range(3)
+    )
+
+
+def auto_import_scale_for_scene(scene):
+    input_extents = []
+    for obj in scene.objects:
+        if obj.type != "EMPTY" or "_MiN_input_scale" not in obj:
+            continue
+        extent_units = _holder_extent_units(obj)
+        if extent_units > 0:
+            input_extents.append(extent_units * float(obj["_MiN_input_scale"]))
+
+    if not input_extents:
+        return None
+
+    return import_scale_for_extent(max(input_extents))
+
+
+def update_import_scale(self, context):
+    # This is essentially a placeholder for a more developed Scene Object that actually knows of its data
+    from ..blender_objects.factories import MinObjectFactory
+    from ..data_model import SceneModel
+    from .min_keys import min_keys
+    from .node_handling import get_min_gn
+
+    if self.MiN_import_scale == AUTO_IMPORT_SCALE:
+        import_scale = auto_import_scale_for_scene(context.scene)
+        if import_scale is not None:
+            self.MiN_import_scale = import_scale
+        return
+
+    scene_model = SceneModel(output_scale=self.MiN_import_scale)
+    for obj in context.scene.objects:
+        if obj.type != "EMPTY" or "_MiN_input_scale" not in obj:
+            continue
+        MinObjectFactory(min_keys.HOLDER, obj=obj).set_scene(scene_model)
+        for child in obj.children:
+            min_gn = get_min_gn(child)
+            if min_gn is None or "axes" not in min_gn.name.lower():
+                continue
+            MinObjectFactory(min_keys.AXES, obj=child).set_scene(scene_model)
+
+
+def import_scale_property(update=update_import_scale):
+    from bpy.props import EnumProperty
+
+    return EnumProperty(
+        name="Microscopy scale -> Blender scale",
+        items=import_scale_items(),
+        description="Defines the scale transform from physical dataset units to Blender meters.",
+        default=AUTO_IMPORT_SCALE,
+        update=update,
+    )
+
+
+def register_import_scale_property(scene_type):
+    scene_type.MiN_import_scale = import_scale_property()
 
 
 def concrete_import_scale_items():
