+++ microscopynodes/load.py
@@ -40,9 +40,7 @@ class Scene():
 
     @property
     def import_scale(self):
-        from .ui.preferences import addon_preferences
-
-        return addon_preferences(bpy.context).import_scale
+        return self.scene.MiN_import_scale
 
     @property
     def output_scale(self):
@@ -56,14 +54,12 @@ class Scene():
 
     def resolve_auto_import_scale(self, dataset_model):
         from .handle_blender_structs.units import AUTO_IMPORT_SCALE, import_scale_for_extent
-        from .ui.preferences import addon_preferences
 
-        prefs = addon_preferences(bpy.context)
-        if prefs.import_scale != AUTO_IMPORT_SCALE:
+        if self.scene.MiN_import_scale != AUTO_IMPORT_SCALE:
             return
         _, _, extent = dataset_model.intermediate_bbox
         input_extent_meters = float(max(extent)) * float(dataset_model.channels[0].data.unit)
-        prefs.import_scale = import_scale_for_extent(input_extent_meters)
+        self.scene.MiN_import_scale = import_scale_for_extent(input_extent_meters)
 
     def update_dataset_scale(self, dataset, dataset_model):
         self.resolve_auto_import_scale(dataset_model)
